[WIP] Add source.hostname/target.hostname to influx output
Grafana is currently not able to resolve the target.id or source.id
into a human readable hostname. Therefore reading the neighbour
graphs is quite difficult for humans. To resolve this, we add the
additional tags source.hostname and target.hostname to the influx
link measurements.

This is just a very first draft, which probably needs some more
var != nil checks or so. It's intended as a basis for discussion.